+++ plugins/rvc/rvc_plugin.py
@@ -1,6 +1,6 @@
 import io
 import re
-from utils import download_and_extract_zip
+from LAV_utils import download_and_extract_zip
 from .inferrvc import load_torchaudio
 from .inferrvc import RVC
 import edge_tts
