+++ subtitle.txt
@@ -1,1 +1,1 @@
-
+すぅ…はぁ——おはようさん、朝の空気は清々しくて気持ちええなぁ、深呼吸して頭もすっきりや。
